pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minervaai"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [
  { name = "Kevin Rohan Vaz", email = "excalibur.krv@gmail.com" },
]
classifiers = [
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "gradio==6.0.1",
  "transformers==4.57.1",
  "diffusers==0.35.2",
  "accelerate==1.11.0",
  "sentencepiece==0.2.1",
  "gguf==0.17.1",
  "ftfy==6.3.1",
  "protobuf==6.33.1",
  "pyarrow==22.0.0",
  "ddgs==9.9.1",
  "ollama==0.6.1",
  "langchain==1.0.7",
  "langchain-ollama==1.0.0",
  "langchain-community==0.4.1",
  "torchvision==0.24.1",
  "modal==1.2.4",
  "kokoro>=0.9.2",
  "soundfile==0.13.1",
  "ffmpeg-python==0.2.0",
  "imageio-ffmpeg==0.6.0",
  "bitsandbytes==0.42.0",
  "langchainhub==0.1.21",
  "numexpr==2.14.1",
  "text-generation==0.7.0",
  "langchain-huggingface==1.1.0",
  "langchain-mcp-adapters==0.1.14",
  "opencv-python==4.12.0.88",
  "imageio"
]

[project.urls]

[tool.hatch.version]
path = "minervaai/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:minervaai tests}"
[tool.hatch.envs.default]
python = "3.12"

[tool.coverage.run]
source_pkgs = ["minervaai", "tests"]
branch = true
parallel = true
omit = [
  "minervaai/__about__.py",
]

[tool.coverage.paths]
minervaai = ["minervaai", "*/minervaai"]
tests = ["tests", "*/minervaai/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
